newstudent fragmnet is done
