+++ app/src/main/AndroidManifest.xml
@@ -13,17 +13,13 @@
         android:theme="@style/Theme.StudentsApp"
         tools:targetApi="31">
         <activity
-            android:name=".StudentsList"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".NewStudent"
-            android:exported="true" >
         </activity>
         <activity
             android:name=".EditStudent"
